+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.6.1 under development
 
+- New #3: Add `StringHelper::truncateWordsByLength()` method to truncate text by character length while preserving word boundaries (@samdark)
 - Bug #119: Fix incorrect pluralization of words ending in "tion", "sion" and "gion" (@samdark)
 
 ## 2.6.0 February 09, 2025